requirements.txt
torch
Pillow
mlx-vlm
PyPDF2
olmocr
torchvision
validators
jiwer
